move sidebar apps and sidebar menu app to same file and extend meta
