+++ libs/extensions-registry/models.ts
@@ -11,8 +11,7 @@ import type { ManifestPropertyAction } from './property-action.models';
 import type { ManifestPropertyEditorUI, ManifestPropertyEditorModel } from './property-editor.models';
 import type { ManifestSection } from './section.models';
 import type { ManifestSectionView } from './section-view.models';
-import type { ManifestSectionSidebarApp } from './section-sidebar-app.models';
-import type { ManifestMenuSectionSidebarApp } from './section-sidebar-menu.models';
+import type { ManifestSectionSidebarApp, ManifestMenuSectionSidebarApp } from './section-sidebar-app.models';
 import type { ManifestMenu } from './menu.models';
 import type { ManifestMenuItem } from './menu-item.models';
 import type { ManifestTheme } from './theme.models';
@@ -40,7 +39,6 @@ export * from './property-editor.models';
 export * from './section-view.models';
 export * from './section.models';
 export * from './section-sidebar-app.models';
-export * from './section-sidebar-menu.models';
 export * from './menu.models';
 export * from './menu-item.models';
 export * from './theme.models';
