+++ libs/extensions-registry/section-sidebar-app.models.ts
@@ -8,3 +8,14 @@ export interface ManifestSectionSidebarApp extends ManifestElement {
 export interface MetaSectionSidebarApp {
 	sections: Array<string>;
 }
+
+// TODO: this is a temp solution until we implement kinds
+export interface ManifestMenuSectionSidebarApp extends ManifestElement {
+	type: 'menuSectionSidebarApp';
+	meta: MetaMenuSectionSidebarApp;
+}
+
+export interface MetaMenuSectionSidebarApp extends MetaSectionSidebarApp {
+	label: string;
+	menu: string;
+}
