--- libs/extensions-registry/section-sidebar-menu.models.ts
@@ -1,12 +1,0 @@
-import type { ManifestElement } from './models';
-
-export interface ManifestMenuSectionSidebarApp extends ManifestElement {
-	type: 'menuSectionSidebarApp';
-	meta: MetaMenuSectionSidebarApp;
-}
-
-export interface MetaMenuSectionSidebarApp {
-	label: string;
-	sections: Array<string>;
-	menu: string;
-}
